use types from harmonia-store-core
We can get rid of nix-libstore this way.

The old nix-libstore and harmonia-store-core store path types
differently, so this required a fair bit of plumbing of the store path.

As part of this, fix NixOS test failures.
Putting packages in environment.systemPackages isn't correct,
it should be done with nix.package. This happened to work with older
nix, but doesn't anymore.

Co-authored-by: Artemis Tosini <artemis.tosini@obsidian.systems>
Co-authored-by: Amaan Qureshi <git@amaanq.com>